Refactor instance() methods to use separate static variables
Add NativeTypeTrait
