+++ src/Phan/Language/Type/CallableArrayType.php
@@ -17,6 +17,8 @@ use Phan\Language\UnionType;
  */
 class CallableArrayType extends ArrayType
 {
+    use NativeTypeTrait;
+
     /** @phan-override */
     public const NAME = 'callable-array';
