+++ src/Phan/Language/Type/FalseType.php
@@ -18,6 +18,8 @@ use Phan\Language\Type;
  */
 final class FalseType extends ScalarType implements LiteralTypeInterface
 {
+    use NativeTypeTrait;
+
     /** @phan-override */
     public const NAME = 'false';
